+++ api/lib/application/organizations-administration/index.js
@@ -1,0 +1,57 @@
+const Joi = require('joi');
+
+const { sendJsonApiError, PayloadTooLargeError } = require('../http-errors.js');
+const securityPreHandlers = require('../security-pre-handlers.js');
+const organizationController = require('./organization-administration-controller.js');
+const identifiersType = require('../../domain/types/identifiers-type.js');
+
+const ERRORS = {
+  PAYLOAD_TOO_LARGE: 'PAYLOAD_TOO_LARGE',
+};
+const TWO_AND_HALF_MEGABYTES = 1048576 * 2.5;
+
+exports.register = async (server) => {
+  server.route([
+    {
+      method: 'PATCH',
+      path: '/api/admin/organizations/{id}',
+      config: {
+        pre: [
+          {
+            method: (request, h) =>
+              securityPreHandlers.adminMemberHasAtLeastOneAccessOf([
+                securityPreHandlers.checkAdminMemberHasRoleSuperAdmin,
+                securityPreHandlers.checkAdminMemberHasRoleSupport,
+                securityPreHandlers.checkAdminMemberHasRoleMetier,
+              ])(request, h),
+            assign: 'hasAuthorizationToAccessAdminScope',
+          },
+        ],
+        validate: {
+          params: Joi.object({
+            id: identifiersType.organizationId,
+          }),
+        },
+        payload: {
+          maxBytes: TWO_AND_HALF_MEGABYTES,
+          failAction: (request, h) => {
+            return sendJsonApiError(
+              new PayloadTooLargeError('An error occurred, payload is too large', ERRORS.PAYLOAD_TOO_LARGE, {
+                maxSizeInMegaBytes: '2.5',
+              }),
+              h
+            );
+          },
+        },
+        handler: organizationController.updateOrganizationInformation,
+        tags: ['api', 'organizations'],
+        notes: [
+          "- **Cette route est restreinte aux utilisateurs authentifiés ayant les droits d'accès**\n" +
+            '- Elle permet de mettre à jour tout ou partie d’une organisation',
+        ],
+      },
+    },
+  ]);
+};
+
+exports.name = 'organizations-administration-api';
