+++ api/lib/application/organizations/index.js
@@ -11,7 +11,6 @@ const ERRORS = {
   PAYLOAD_TOO_LARGE: 'PAYLOAD_TOO_LARGE',
 };
 const TWENTY_MEGABYTES = 1048576 * 20;
-const TWO_AND_HALF_MEGABYTES = 1048576 * 2.5;
 
 exports.register = async (server) => {
   const adminRoutes = [
@@ -165,45 +164,6 @@ exports.register = async (server) => {
         notes: [
           "- **Cette route est restreinte aux utilisateurs authentifiés ayant les droits d'accès**\n" +
             "- Elle permet d'archiver une organisation",
-        ],
-      },
-    },
-    {
-      method: 'PATCH',
-      path: '/api/admin/organizations/{id}',
-      config: {
-        pre: [
-          {
-            method: (request, h) =>
-              securityPreHandlers.adminMemberHasAtLeastOneAccessOf([
-                securityPreHandlers.checkAdminMemberHasRoleSuperAdmin,
-                securityPreHandlers.checkAdminMemberHasRoleSupport,
-                securityPreHandlers.checkAdminMemberHasRoleMetier,
-              ])(request, h),
-            assign: 'hasAuthorizationToAccessAdminScope',
-          },
-        ],
-        validate: {
-          params: Joi.object({
-            id: identifiersType.organizationId,
-          }),
-        },
-        payload: {
-          maxBytes: TWO_AND_HALF_MEGABYTES,
-          failAction: (request, h) => {
-            return sendJsonApiError(
-              new PayloadTooLargeError('An error occurred, payload is too large', ERRORS.PAYLOAD_TOO_LARGE, {
-                maxSizeInMegaBytes: '2.5',
-              }),
-              h
-            );
-          },
-        },
-        handler: organizationController.updateOrganizationInformation,
-        tags: ['api', 'organizations'],
-        notes: [
-          "- **Cette route est restreinte aux utilisateurs authentifiés ayant les droits d'accès**\n" +
-            '- Elle permet de mettre à jour tout ou partie d’une organisation',
         ],
       },
     },
