+++ api/lib/domain/usecases/index.js
@@ -672,7 +672,7 @@ const updateCertificationCenterReferer = require('./update-certification-center-
 const updateExpiredPassword = require('./update-expired-password.js');
 const updateLastQuestionState = require('./update-last-question-state.js');
 const updateMembership = require('./update-membership.js');
-const updateOrganizationInformation = require('./update-organization-information.js');
+const updateOrganizationInformation = require('./organizations-administration/update-organization.js');
 const updateOrganizationLearnerDependentUserPassword = require('./update-organization-learner-dependent-user-password.js');
 const updateParticipantExternalId = require('./update-participant-external-id.js');
 const updateSession = require('./update-session.js');
