+++ api/tests/acceptance/application/organizations-administration/_files/logo-3mb.js
